Function for Finding Element in tuple
This Function finds an element in a tuple; if it is found, it gives an index of the element as output. 
This function can handle nested objects like list, tuple, dictionary, and set within a tuple, but it will only return the index of the main element in the case of a set and dictionary.